@@ -8,7 +8,7 @@ Gem::Specification.new do |s|
   s.email = "smartinez87@gmail.com"
   s.license = "MIT"
 
-  s.required_ruby_version     = '>= 2.0'
+  s.required_ruby_version     = '>= 2.2'
   s.required_rubygems_version = '>= 1.8.11'
 
   s.files = `git ls-files`.split("\n")
@@ -16,10 +16,10 @@ Gem::Specification.new do |s|
   s.test_files = `git ls-files -- test`.split("\n")
   s.require_path = 'lib'
 
-  s.add_dependency("actionmailer", "~> 4.0")
-  s.add_dependency("activesupport", "~> 4.0")
+  s.add_dependency("actionmailer", "5.0.0.beta1")
+  s.add_dependency("activesupport", "5.0.0.beta1")
 
-  s.add_development_dependency "rails", "~> 4.0"
+  s.add_development_dependency "rails", "5.0.0.beta1"
   s.add_development_dependency "resque", "~> 1.2.0"
   # Sidekiq 3.2.2 does not support Ruby 1.9. 
   s.add_development_dependency "sidekiq", "~> 3.0.0", "< 3.2.2"
