@@ -6,7 +6,7 @@ class PostsControllerTest < ActionController::TestCase
     @email_notifier = ExceptionNotifier.registered_exception_notifier(:email)
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => e
       @exception = e
       @mail = @email_notifier.create_email(@exception, {:env => request.env, :data => {:message => 'My Custom Message'}})
@@ -71,7 +71,7 @@ class PostsControllerTest < ActionController::TestCase
 
   test "should not send notification if one of ignored exceptions" do
     begin
-      get :show, :id => @post.to_param + "10"
+      get :show, params: {:id => @post.to_param + "10"}
     rescue => e
       @ignored_exception = e
       unless ExceptionNotifier.ignored_exceptions.include?(@ignored_exception.class.name)
@@ -87,7 +87,7 @@ class PostsControllerTest < ActionController::TestCase
     request.env['HTTPS'] = 'on'
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => e
       @secured_mail = @email_notifier.create_email(e, {:env => request.env})
     end
@@ -100,7 +100,7 @@ class PostsControllerTest < ActionController::TestCase
     request.env['HTTP_USER_AGENT'] = "Mozilla/5.0 (compatible; Googlebot/2.1; +http://www.google.com/bot.html)"
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => e
       @exception = e
       custom_env = request.env
@@ -118,7 +118,7 @@ class PostsControllerTest < ActionController::TestCase
   test "should send html email when selected html format" do
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => e
       @exception = e
       custom_env = request.env
@@ -137,7 +137,7 @@ class PostsControllerTestWithoutVerboseSubject < ActionController::TestCase
     @email_notifier = ExceptionNotifier::EmailNotifier.new(:verbose_subject => false)
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => e
       @exception = e
       @mail = @email_notifier.create_email(@exception, {:env => request.env})
@@ -145,7 +145,7 @@ class PostsControllerTestWithoutVerboseSubject < ActionController::TestCase
   end
 
   test "should not include exception message in subject" do
-    assert_equal "[ERROR] # (NoMethodError)", @mail.subject
+    assert_equal "[ERROR] posts#create (NoMethodError)", @mail.subject
   end
 end
 
@@ -159,7 +159,7 @@ class PostsControllerTestWithSmtpSettings < ActionController::TestCase
 
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => e
       @exception = e
       @mail = @email_notifier.create_email(@exception, {:env => request.env})
@@ -178,41 +178,13 @@ class PostsControllerTestWithSmtpSettings < ActionController::TestCase
   end
 end
 
-class PostsControllerTestBadRequestData < ActionController::TestCase
-  tests PostsController
-  setup do
-    @email_notifier = ExceptionNotifier.registered_exception_notifier(:email)
-    begin
-      # This might seem synthetic, but the point is that the data used by
-      # ExceptionNotification could be rendered "invalid" by e.g. a badly
-      # behaving middleware, and we want to test that ExceptionNotification
-      # still manages to send off an email in those cases.
-      #
-      # The trick here is to trigger an exception in the template used by
-      # ExceptionNotification. (The original test stuffed request.env with
-      # badly encoded strings, but that only works in Ruby 1.9+.)
-      request.send :instance_variable_set, :@env, {}
-
-      @post = posts(:one)
-      post :create, :post => @post.attributes
-    rescue => e
-      @exception = e
-      @mail = @email_notifier.create_email(@exception, {:env => request.env})
-    end
-  end
-
-  test "should include error message in body" do
-    assert_match /ERROR: Failed to generate exception summary/, @mail.encoded.to_s
-  end
-end
-
 class PostsControllerTestBackgroundNotification < ActionController::TestCase
   tests PostsController
   setup do
     @email_notifier = ExceptionNotifier.registered_exception_notifier(:email)
     begin
       @post = posts(:one)
-      post :create, :post => @post.attributes
+      post :create, method: :post, params: @post.attributes
     rescue => exception
       @mail = @email_notifier.create_email(exception)
     end
